+++ base/src/main/kotlin/com/freeletics/gradle/setup/PaparazziSetup.kt
@@ -1,7 +1,0 @@
-package com.freeletics.gradle.setup
-
-import org.gradle.api.Project
-
-internal fun Project.configurePaparazzi() {
-    plugins.apply("com.freeletics.fork.paparazzi")
-}