+++ monorepo/src/main/kotlin/com/freeletics/gradle/plugin/LegacyExtension.kt
@@ -1,5 +1,0 @@
-package com.freeletics.gradle.plugin
-
-public abstract class LegacyExtension {
-    public var allowLegacyDependencies: Boolean = false
-}