+++ plugins/src/main/kotlin/com/freeletics/gradle/monorepo/plugin/CoreKotlinPlugin.kt
@@ -1,0 +1,35 @@
+package com.freeletics.gradle.monorepo.plugin
+
+import com.freeletics.gradle.monorepo.setup.addDefaultDependencies
+import com.freeletics.gradle.monorepo.setup.addTestDependencies
+import com.freeletics.gradle.monorepo.tasks.CheckDependencyRulesTask.Companion.registerCheckDependencyRulesTasks
+import com.freeletics.gradle.monorepo.util.ProjectType
+import com.freeletics.gradle.plugin.FreeleticsJvmPlugin
+import com.freeletics.gradle.util.freeleticsJvmExtension
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+public abstract class CoreKotlinPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.plugins.apply(FreeleticsJvmPlugin::class.java)
+
+        target.freeleticsJvmExtension.useAndroidLint()
+
+        target.dependencies.apply {
+            addDefaultDependencies(target)
+            addTestDependencies(target)
+        }
+
+        target.registerCheckDependencyRulesTasks(
+            allowedProjectTypes = listOf(
+                ProjectType.CORE_API,
+                ProjectType.CORE_IMPLEMENTATION,
+                ProjectType.CORE_TESTING,
+            ),
+            allowedDependencyProjectTypes = listOfNotNull(
+                ProjectType.CORE_API,
+                ProjectType.CORE_TESTING,
+            ),
+        )
+    }
+}