+++ plugins/src/main/kotlin/com/freeletics/gradle/monorepo/plugin/DomainAndroidPlugin.kt
@@ -1,0 +1,54 @@
+package com.freeletics.gradle.monorepo.plugin
+
+import com.freeletics.gradle.monorepo.setup.addAndroidDependencies
+import com.freeletics.gradle.monorepo.setup.addDefaultDependencies
+import com.freeletics.gradle.monorepo.setup.addTestDependencies
+import com.freeletics.gradle.monorepo.tasks.CheckDependencyRulesTask.Companion.registerCheckDependencyRulesTasks
+import com.freeletics.gradle.monorepo.util.ProjectType
+import com.freeletics.gradle.monorepo.util.appType
+import com.freeletics.gradle.monorepo.util.projectType
+import com.freeletics.gradle.plugin.FreeleticsAndroidPlugin
+import com.freeletics.gradle.util.freeleticsAndroidExtension
+import com.freeletics.gradle.util.freeleticsExtension
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+
+public abstract class DomainAndroidPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.plugins.apply(FreeleticsAndroidPlugin::class.java)
+
+        val extension = target.freeleticsExtension.extensions.create("legacy", LegacyExtension::class.java)
+
+        target.freeleticsAndroidExtension.minSdkVersion(target.appType()?.minSdkVersion(target))
+        target.freeleticsAndroidExtension.enableParcelize()
+
+        target.dependencies.apply {
+            addDefaultDependencies(target)
+            addAndroidDependencies(target)
+            addTestDependencies(target)
+        }
+
+        target.afterEvaluate {
+            target.registerCheckDependencyRulesTasks(
+                allowedProjectTypes = listOf(
+                    ProjectType.DOMAIN_API,
+                    ProjectType.DOMAIN_IMPLEMENTATION,
+                    ProjectType.DOMAIN_TESTING,
+                ),
+                allowedDependencyProjectTypes = listOfNotNull(
+                    ProjectType.CORE_API,
+                    ProjectType.CORE_TESTING,
+                    ProjectType.DOMAIN_API,
+                    ProjectType.DOMAIN_TESTING,
+                    if (target.projectType() == ProjectType.DOMAIN_IMPLEMENTATION) ProjectType.FEATURE_NAV else null,
+                    // TODO remove after migrating domain modules away from legacy
+                    if (extension.allowLegacyDependencies) {
+                        ProjectType.LEGACY
+                    } else {
+                        null
+                    },
+                ),
+            )
+        }
+    }
+}